requirements.txt
Flask
Flask-Cors
g4f
werkzeug
psycopg2-binary
gunicorn
python-dotenv
psycopg2